backend/requirements.txt
Flask
opencv-python
tensorflow
tensorflow-hub
flask-cors
pillow